ref: DialogForm moved to dialogs folder and Spinner added for loading states.
