+++ src/components/not-found.tsx
@@ -1,17 +1,29 @@
 import { Link, useNavigate, useRouteError } from "react-router-dom";
 import { Button } from "./ui/button";
-
+const errorMessage = (error: unknown) => {
+  if (error instanceof Error) {
+    return error.message;
+  }
+  else if (error && typeof error === 'object' && 'statusText' in error) {
+    return error.statusText as string;
+  }
+  else if (typeof error === 'string') {
+    return error;
+  }
+  else {
+    return 'Unknown error';
+  }
+}
 export default function NotFound() {
   const error = useRouteError();
   console.error(error);
   const navigate = useNavigate();
   return (
-    <main className="w-full h-screen flex flex-col items-center justify-center pb-12">
-      <div className="text-destructive-foreground">
+    <main className="w-full h-[calc(100vh-60px)] flex flex-col items-center justify-center pb-12 text-center">
+      <div className="text-destructive-foreground p-6 space-y-4">
         <h1>Oops!</h1>
-        <p>Sorry, an unexpected error has occurred.</p>
+        <p className="">Sorry, an unexpected error has occurred.</p>
         <p>
-          {/* <i>{error.statusText || error.message}</i> */}
           <i>
             {errorMessage(error)}
           </i>
@@ -30,19 +42,4 @@ export default function NotFound() {
       </div>
     </main>
   );
-}
-
-const errorMessage = (error: unknown) => {
-  if (error instanceof Error) {
-    return error.message;
-  }
-  else if (error && typeof error === 'object' && 'statusText' in error) {
-    return error.statusText as string;
-  }
-  else if (typeof error === 'string') {
-    return error;
-  }
-  else {
-    return 'Unknown error';
-  }
 }
